+++ tensorflow_quantum/python/layers/circuit_construction/unsigned_test.py
@@ -12,254 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Tests for the qudit layers."""
-import tensorflow as tf
-import cirq
-import sympy
-
-from tensorflow_quantum.python import util
-from tensorflow_quantum.python.layers.circuit_construction import unsigned
-from tensorflow_quantum.python.layers.circuit_executors import expectation
-
-
-def _append_register_bits(r, r_int, precisions, register_list):
-    """Generate a bitstring corresponding to r_int on the indicated register."""
-    r_int_str = format(r_int, 'b')
-    bit_circ = cirq.Circuit()
-    for n, c in enumerate(r_int_str[::-1]):
-        if bool(int(c)):
-            bit_circ += cirq.X(register_list[r][precisions[r] - 1 - n])
-    for q in register_list[r]:
-        bit_circ += cirq.I(q)
-    return bit_circ
-
-
-class ProjectorOnOneTest(tf.test.TestCase):
-    """Test the projector_on_one function."""
-
-    def test_bad_inputs(self):
-        """Confirm that function raises error on bad input."""
-        for bad_arg in [-5, 12, cirq.LineQubit(6), []]:
-            with self.assertRaisesRegex(TypeError,
-                                        expected_regex="cirq.GridQubit"):
-                _ = unsigned.projector_on_one(bad_arg)
-
-    def test_return(self):
-        """Confirm that GridQubit input returns correct PauliSum."""
-        for i in range(12):
-            for j in range(12):
-                this_q = cirq.GridQubit(i, j)
-                expected_psum = 0.5 * cirq.I(this_q) - 0.5 * cirq.Z(this_q)
-                test_psum = unsigned.projector_on_one(this_q)
-                self.assertEqual(expected_psum, test_psum)
-
-
-class IntegerOperatorTest(tf.test.TestCase):
-    """Test the integer_operator function."""
-
-    def test_bad_inputs(self):
-        """Confirm that function raises error on bad input."""
-        for bad_arg in [-7, 3, cirq.LineQubit(2), ["junk"]]:
-            with self.assertRaisesRegex(TypeError,
-                                        expected_regex="cirq.GridQubit"):
-                _ = unsigned.integer_operator(bad_arg)
-
-    def test_return(self):
-        """Confirm that correct operators are created."""
-        for i in range(5):
-            qubits = cirq.GridQubit.rect(1, i + 1)
-            expected_psum = (2**(i + 1) - 1) * 0.5 * cirq.I(qubits[0])
-            for loc, q in enumerate(qubits):
-                expected_psum -= 2**(i - loc) * 0.5 * cirq.Z(q)
-            test_psum = unsigned.integer_operator(qubits)
-            self.assertEqual(expected_psum, test_psum)
-
-
-class RegistersFromPrecisionsTest(tf.test.TestCase):
-    """Test the registers_from_precisions function."""
-
-    def test_bad_inputs(self):
-        """Confirm that function raises error on bad input."""
-        for bad_arg in [-7, 3, cirq.LineQubit(2), [[]], ["junk"]]:
-            with self.assertRaisesRegex(TypeError, expected_regex=""):
-                _ = unsigned.registers_from_precisions(bad_arg)
-
-    def test_return(self):
-        """Confirm that registers are built correctly."""
-        precisions = [3, 2, 4]
-        register_list = unsigned.registers_from_precisions(precisions)
-        test_set = set()
-        for r, p in zip(register_list, precisions):
-            self.assertEqual(len(r), p)
-            for q in r:
-                test_set.add(q)
-                self.assertIsInstance(q, cirq.GridQubit)
-        # Ensure all qubits are unique
-        self.assertEqual(len(test_set), sum(precisions))
-
-
-class BuildCliquesPsumTest(tf.test.TestCase):
-    """Test the build_cliques_psum function."""
-
-    def test_bad_inputs(self):
-        """Confirm function raises error on bad inputs."""
-        p = [3, 4]
-        c = {(0,): 2, (1,): 3, (0, 1): 4}
-        with self.assertRaisesRegex(TypeError, expected_regex="list"):
-            _ = unsigned.build_cliques_psum("junk", c)
-        with self.assertRaisesRegex(TypeError, expected_regex="integer"):
-            _ = unsigned.build_cliques_psum(["junk"], c)
-        with self.assertRaisesRegex(TypeError, expected_regex="greater than"):
-            _ = unsigned.build_cliques_psum([2, 0], c)
-        with self.assertRaisesRegex(ValueError, expected_regex="Cannot access"):
-            # All labels must be in range
-            _ = unsigned.build_cliques_psum(p, {(0, 1, 2): 3})
-        with self.assertRaisesRegex(TypeError, expected_regex="dict"):
-            _ = unsigned.build_cliques_psum(p, "junk")
-        with self.assertRaisesRegex(TypeError, expected_regex="dict"):
-            _ = unsigned.build_cliques_psum(p, [1])
-        with self.assertRaisesRegex(TypeError, expected_regex="tuple"):
-            # Most common error, where the key is not parsed as a tuple
-            _ = unsigned.build_cliques_psum(p, {(0): 1})
-        with self.assertRaisesRegex(TypeError, expected_regex="tuple"):
-            # Most common error, where the key is not parsed as a tuple
-            _ = unsigned.build_cliques_psum(p, {(0,): 1, (0): 1})
-        with self.assertRaisesRegex(TypeError, expected_regex="tuple"):
-            _ = unsigned.build_cliques_psum(p, {"junk": 1})
-        with self.assertRaisesRegex(TypeError, expected_regex="integer"):
-            _ = unsigned.build_cliques_psum(p, {(0, 1, "junk"): 1})
-        with self.assertRaisesRegex(TypeError, expected_regex="non-negative"):
-            _ = unsigned.build_cliques_psum(p, {(0, 1, -1): 1})
-        with self.assertRaisesRegex(TypeError, expected_regex="integer"):
-            _ = unsigned.build_cliques_psum(p, {(2.7,): 1})
-        with self.assertRaisesRegex(TypeError, expected_regex="value"):
-            _ = unsigned.build_cliques_psum(p, {(0,): "junk"})
-        with self.assertRaisesRegex(TypeError, expected_regex="value"):
-            _ = unsigned.build_cliques_psum(p, {(0,): 1, (1,): 1j})
-
-    def test_psum_return(self):
-        """Confirm operators are built correctly."""
-        p = [3, 4]
-        c = {(0,): 2, (1,): 3, (0, 1): 4, (): 7}
-        test_psum = unsigned.build_cliques_psum(p, c)
-        registers = unsigned.registers_from_precisions(p)
-        j0 = unsigned.integer_operator(registers[0])
-        j1 = unsigned.integer_operator(registers[1])
-        expected_psum = 2 * j0 + 3 * j1 + 4 * j0 * j1 + 7 * cirq.PauliString(
-            cirq.I(registers[0][0]))
-        self.assertEqual(expected_psum, test_psum)
-
-    def test_counting(self):
-        """Confirm we get numbers as expected from quantum integer psums."""
-        precisions = [5, 3, 4]
-        registers = unsigned.registers_from_precisions(precisions)
-        cliques = [{(0,): 1}, {(1,): 1}, {(2,): 1}]
-        cliques_psums = [
-            unsigned.build_cliques_psum(precisions, c) for c in cliques
-        ]
-        exp_layer = expectation.Expectation()
-        # Test that all counts increment correctly
-        for i in range(2**precisions[0]):
-            bit_circ_i = _append_register_bits(0, i, precisions, registers)
-            for j in range(2**precisions[1]):
-                bit_circ_j = _append_register_bits(1, j, precisions, registers)
-                for k in range(2**precisions[2]):
-                    bit_circ_k = _append_register_bits(2, k, precisions,
-                                                       registers)
-                    test_val = exp_layer(bit_circ_i + bit_circ_j + bit_circ_k,
-                                         operators=cliques_psums)
-                    self.assertAllClose([[i, j, k]],
-                                        test_val.numpy(),
-                                        atol=1e-5)
-
-    def test_squares(self):
-        """Confirm we get squared numbers as expected from q integer psums."""
-        precisions = [5]
-        registers = unsigned.registers_from_precisions(precisions)
-        cliques = {(0, 0): 1}
-        cliques_psums = unsigned.build_cliques_psum(precisions, cliques)
-        exp_layer = expectation.Expectation()
-        # Test that all counts increment correctly
-        for i in range(2**precisions[0]):
-            bit_circ_i = _append_register_bits(0, i, precisions, registers)
-            test_val = exp_layer(bit_circ_i, operators=cliques_psums)
-            self.assertAllClose([[i**2]], test_val.numpy(), atol=1e-5)
-
-    def test_polynomial(self):
-        """Confirm that expectations of polynomial with offset are correct."""
-        precisions = [5]
-        registers = unsigned.registers_from_precisions(precisions)
-        # test the polynomial y = 3 + 5.2x - 7.5x**2 + 0.3x**3
-        cliques = {(): 3, (0,): 5.2, (0, 0): -7.5, (0, 0, 0): 0.3}
-        cliques_psums = unsigned.build_cliques_psum(precisions, cliques)
-        exp_layer = expectation.Expectation()
-        # Test that all counts increment correctly
-        for i in range(2**precisions[0]):
-            bit_circ_i = _append_register_bits(0, i, precisions, registers)
-            test_val = exp_layer(bit_circ_i, operators=cliques_psums)
-            self.assertAllClose([[3 + 5.2 * i - 7.5 * i**2 + 0.3 * i**3]],
-                                test_val.numpy(),
-                                atol=1e-4)
-
-
-class BuildCliquesExpTest(tf.test.TestCase):
-    """Test the build_cliques_exp function."""
-
-    def test_build_cliques_exp_error(self):
-        """Test that build_cliques_exp raises error on bad input."""
-        p = [5]
-        c = {(0,): 1}
-        with self.assertRaisesRegex(TypeError, expected_regex=""):
-            _ = unsigned.build_cliques_exp("junk", c)
-        with self.assertRaisesRegex(TypeError, expected_regex=""):
-            _ = unsigned.build_cliques_exp(p, "junk")
-        with self.assertRaisesRegex(TypeError, expected_regex=""):
-            _ = unsigned.build_cliques_exp(p, c, [])
-
-    def test_build_cliques_exp_return(self):
-        """Test that the correct circuit is generated."""
-        p = [3, 4]
-        c = {(1,): 2}
-        for coeff in [3.4, "gamma", sympy.Symbol("symbol")]:
-            expected_circuit = util.exponential(
-                [unsigned.build_cliques_psum(p, c)], [coeff])
-            test_circuit = unsigned.build_cliques_exp(p, c, coeff)
-            self.assertTrue(
-                cirq.protocols.approx_eq(expected_circuit, test_circuit))
-
-
-class BuildMomentaExpTest(tf.test.TestCase):
-    """Test the build_mometa_exp function."""
-
-    def test_build_momenta_exp_layer_error(self):
-        """Test that build_momenta_exp raises error on bad input."""
-        p = [5]
-        c = {(0,): 1}
-        with self.assertRaisesRegex(TypeError, expected_regex=""):
-            _ = unsigned.build_momenta_exp("junk", c)
-        with self.assertRaisesRegex(TypeError, expected_regex=""):
-            _ = unsigned.build_momenta_exp(p, "junk")
-        with self.assertRaisesRegex(TypeError, expected_regex=""):
-            _ = unsigned.build_momenta_exp(p, c, [])
-
-    def test_build_momenta_exp_return(self):
-        """Test that the correct circuit is generated."""
-        p = [3, 4]
-        c = {(1,): 2}
-        r = unsigned.registers_from_precisions(p)
-        qft_0 = cirq.Circuit(cirq.QFT(*r[0]))
-        qft_1 = cirq.Circuit(cirq.QFT(*r[1]))
-        convert = cirq.ConvertToCzAndSingleGates(allow_partial_czs=True)
-        convert(qft_0)
-        convert(qft_1)
-        for coeff in [3.4, "gamma", sympy.Symbol("symbol")]:
-            expected_circuit = (
-                qft_0 + qft_1 +
-                util.exponential([unsigned.build_cliques_psum(p, c)], [coeff]) +
-                qft_1**-1 + qft_0**-1)
-            test_circuit = unsigned.build_momenta_exp(p, c, coeff)
-            self.assertTrue(
-                cirq.protocols.approx_eq(expected_circuit, test_circuit))
 
 
 if __name__ == "__main__":
